Use keepalive for all components
This specifically solves the issue on the checkbox view where the 'All'
value was set to true every time a user navigated to that view.
This also introduced the issue of the basket view not updating stats
because the component didn't mount again on every render. This is
addressed in a follow-up commit.