Do not shallow copy filter arrays or all arrays
This addresses the issue where the values in a basket item would change
when a user navigated away from the basket, then updated selections,
then created a new basket item and navigated back to the basket. The
basket items seemed to not have been created with independent arrays.
This commit changes that.